+++ changelog.md
@@ -1,2 +1,5 @@
 ## Paper
-- fix bug that broke all connected break-chained blocks, even vertical crops
+
+- do not use magnetic's break-chained block logic if the player is not eligible to use magnetic
+- fixes bug with break-chained blocks where some items were simply not dropped if the first-broken block did not also
+  drop it